Optimize max expected pressure

+++ 2022/day16/day16_1.cpp
@@ -242,7 +242,7 @@ struct SystemEq
     }
 };
 
-using Pressures = std::unordered_map<System, size_t, SystemHash, SystemEq>;
+using Pressures = std::unordered_map<System, uint16_t, SystemHash, SystemEq>;
 
 static Pressures dijkstra()
 {
